# api/openapi.yaml
openapi: 3.0.3
info:
  title: Modura HTTP API
  version: 0.1.0
  description: Authoritative HTTP contract for the Modura modular monolith.
servers:
  - url: /api
paths:
...
  /platform/tenants:
    get:
      operationId: listPlatformTenants
      summary: List tenants as a global platform administrator
      tags: [platform-tenants]
      security:
        - platformBearerAuth: []
      responses:
        "200":
          description: Tenant summaries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlatformTenant"
        "401":
          $ref: "#/components/responses/AuthenticationFailed"
  /platform/tenants/{tenantId}/suspend:
    post:
      operationId: suspendPlatformTenant
      summary: Suspend an active tenant with auditable reason
      tags: [platform-tenants]
      security:
        - platformBearerAuth: []
      parameters:
        - $ref: "#/components/parameters/TenantId"
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TenantLifecycleRequest"
      responses:
        "204":
          description: Tenant suspended and audit evidence committed
        "400":
          description: Invalid lifecycle transition
        "401":
          $ref: "#/components/responses/AuthenticationFailed"
        "403":
          $ref: "#/components/responses/CsrfFailed"
        "404":
          description: Tenant not found
  /platform/tenants/{tenantId}/reactivate:
    post:
      operationId: reactivatePlatformTenant
      summary: Reactivate a suspended tenant with auditable reason
      tags: [platform-tenants]
      security:
        - platformBearerAuth: []
      parameters:
        - $ref: "#/components/parameters/TenantId"
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TenantLifecycleRequest"
      responses:
        "204":
          description: Tenant reactivated and audit evidence committed
        "400":
          description: Invalid lifecycle transition
        "401":
          $ref: "#/components/responses/AuthenticationFailed"
        "403":
          $ref: "#/components/responses/CsrfFailed"
        "404":
          description: Tenant not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    platformBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema:
        type: string
        minLength: 32
        maxLength: 256
...
    TenantId:
      name: tenantId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    AuthenticationFailed:
      description: Tenant or credentials are invalid, or the session is unavailable
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    CsrfFailed:
      description: CSRF validation failed
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
  schemas:
...
    PlatformTenant:
      type: object
      additionalProperties: false
      required: [id, slug, displayName, status, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
        displayName:
          type: string
        status:
          type: string
          enum: [provisioning, active, suspended]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    TenantLifecycleRequest:
      type: object
      additionalProperties: false
      required: [reason]
      properties:
        reason:
          type: string
          minLength: 1
          maxLength: 500
...
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        traceId:
          type: string